pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pr-assistant"
version = "0.1.0"
description = "AI-first PR Assistant service"
readme = "AGENTS.md"
requires-python = ">=3.13"
dependencies = [
  "fastapi>=0.116.0,<1.0.0",
  "httpx>=0.28.0,<1.0.0",
  "pydantic-settings>=2.10.0,<3.0.0",
  "PyJWT[crypto]>=2.10.0,<3.0.0",
  "uvicorn>=0.35.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.0,<9.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
